requirements.txt
fastapi==0.75.0
uvicorn==0.22.0
tensorflow-gpu==2.4.0
transformers==4.0.0